Move ProductionPage to the pages/ directory of features/archive/
