+++ frontend/app/features/archive/pages/ProductionPage.tsx
@@ -11,8 +11,8 @@ import type { Event, Price } from "~/features/archive/types/eventTypes";
 import { useLocalizedPath } from "~/shared/hooks/useLocalizedPath";
 import { getEventByUrl, getPriceByUrl } from "~/features/archive/services/eventService";
 import { getHallByUrl } from "~/features/archive/services/hallService";
-import { EventCard, type EventWithResolvedRelations } from "./EventCard";
-import { ProductionPageMediaGallery } from "./ProductionPageMediaGallery";
+import { EventCard, type EventWithResolvedRelations } from "../components/EventCard";
+import { ProductionPageMediaGallery } from "../components/ProductionPageMediaGallery";
 import { Protected } from "~/features/auth";
 import { ARCHIVE_PERMISSIONS } from "../archive.constants";
 import { updateProductionByUrl } from "../services/productionService";
